Refactor ReadingGoalForm
