+++ src/main/java/com/karankumar/bookproject/ui/goal/ReadingGoalForm.java
@@ -166,29 +166,4 @@ public class ReadingGoalForm extends VerticalLayout {
      * Vaadin's event bus system. A registered listener can be notified when a save or delete event
      * is fired
      */
-    public abstract static class GoalFormEvent extends ComponentEvent<ReadingGoalForm> {
-        private final ReadingGoal readingGoal;
-
-        protected GoalFormEvent(ReadingGoalForm source, ReadingGoal readingGoal) {
-            super(source, false);
-            this.readingGoal = readingGoal;
-        }
-
-        public ReadingGoal getReadingGoal() {
-            return readingGoal;
-        }
-    }
-
-    public static class SaveEvent extends GoalFormEvent {
-        SaveEvent(ReadingGoalForm source, ReadingGoal readingGoal) {
-            super(source, readingGoal);
-        }
-    }
-
-    // TODO: implement deleting reading goal
-    public static class DeleteEvent extends GoalFormEvent {
-        DeleteEvent(ReadingGoalForm source, ReadingGoal readingGoal) {
-            super(source, readingGoal);
-        }
-    }
 }
